# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Hology 8 API Documentation",
        "title": "Hology 8 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.hology.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/provinces": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetch all available provinces data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provinces"
                ],
                "summary": "Fetch all provinces data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.ProvinceResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/provinces/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetch one province data by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Provinces"
                ],
                "summary": "Fetch provinces data by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Province ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ProvinceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Province with requested ID not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ProvinceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "province_name": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/response.Status"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/response.Status"
                }
            }
        },
        "response.Status": {
            "type": "string",
            "enum": [
                "success",
                "fail",
                "error"
            ],
            "x-enum-varnames": [
                "Success",
                "Fail",
                "Error"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for accessing all endpoints. Type: Key TOKEN",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        },
        "UserAuth": {
            "description": "API Key for accessing protected user and admin endpoints. Type: Bearer TOKEN",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}